refactor(ui): extract event logic of plugin kit out
